Separate responder and middleware from handlers.
