+++ pkg/web/handlers/handler.go
@@ -4,6 +4,7 @@ import (
 	"net/http"
 
 	"github.com/arskydev/weatherman/pkg/service"
+	"github.com/arskydev/weatherman/pkg/web/internal/middleware"
 	"github.com/gorilla/mux"
 )
 
@@ -19,12 +20,13 @@ func NewHandler(s *service.Service) *Handler {
 }
 
 func (h *Handler) InitRoutes() http.Handler {
+	m := middleware.New(h.service.Authorization)
 	r := mux.NewRouter()
 	// AUTH
 	r.HandleFunc("/auth/sign-up", h.signUp).Methods("POST")
 	r.HandleFunc("/auth/sign-in", h.signIn).Methods("POST")
 	// API
-	r.Handle("/api/get-mock", h.ValidateJWT(h.getMock)).Methods("GET")
-	r.Handle("/api/get-weather", h.ValidateJWT(h.getWeather)).Methods("GET")
+	r.Handle("/api/get-mock", m.ValidateJWT(h.getMock)).Methods("GET")
+	r.Handle("/api/get-weather", m.ValidateJWT(h.getWeather)).Methods("GET")
 	return r
 }
