+++ pkg/web/internal/middleware/middleware.go
@@ -1,0 +1,13 @@
+package middleware
+
+import "github.com/arskydev/weatherman/pkg/service"
+
+type Middleware struct {
+	auth service.Authorization
+}
+
+func New(auth service.Authorization) *Middleware {
+	return &Middleware{
+		auth: auth,
+	}
+}
