+++ pkg/web/internal/middleware/validate_jwt.go
@@ -1,29 +1,31 @@
-package handlers
+package middleware
 
 import (
 	"errors"
 	"net/http"
+
+	"github.com/arskydev/weatherman/pkg/web/internal/responder"
 )
 
 const (
 	AUTH_HEADER_NAME = "Authorization"
 )
 
-func (h *Handler) ValidateJWT(next func(http.ResponseWriter, *http.Request)) http.Handler {
+func (m *Middleware) ValidateJWT(next func(http.ResponseWriter, *http.Request)) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		w.Header().Set("Connection", "close")
 		defer r.Body.Close()
 		if r.Header[AUTH_HEADER_NAME] == nil {
 			msg := "empty token"
-			h.sendErrorResponse(msg, http.StatusUnauthorized, w, errors.New(msg))
+			responder.SendErrorResponse(msg, http.StatusUnauthorized, w, errors.New(msg))
 			return
 		}
 
-		token, err := h.service.ValidateToken(r.Header[AUTH_HEADER_NAME][0])
+		token, err := m.auth.ValidateToken(r.Header[AUTH_HEADER_NAME][0])
 
 		if err != nil {
 			msg := "Access denied"
-			h.sendErrorResponse(msg, http.StatusForbidden, w, err)
+			responder.SendErrorResponse(msg, http.StatusForbidden, w, err)
 			return
 		}
